+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.9.3
+
+* Updates `androidx.media3` to 1.9.1.
+
 ## 2.9.2
 
 * Bumps kotlin_version to 2.3.0.