+++ packages/video_player/video_player_avfoundation/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.9.3
+
+* Fixes a regression where HTTP headers were ignored.
+
+## 2.9.2
+
+* Refactors for improved testability.
+
+## 2.9.1
+
+* Refactors native code for improved testability.
+
 ## 2.9.0
 
 * Implements `getAudioTracks()` and `selectAudioTrack()` methods.